NGC-2463 Fix memory leak
Using a new ActorMaterializer instead of the Play app's Materializer seems to cause a memory leak in native-apps-api-orchestration (this can be reproduced by running the performance tests in ngc-performance-tests).

This class seems to have been copied from frontend-bootstrap, and before the Play 2.5 upgrade it did not override the Materializer. frontend-bootstrap's Play 2.5 version does not override the Materializer. This suggests the override is not necessary.